include/ofi_hmem: move IPC defines to ofi.h
Move IPC definitions from ofi_hmem.h to ofi.h to put them in a
more common location
This allows us to remove the include of ofi_hmem.h in ofi_mr.h

Add include of ofi_hmem.h in other places as needed

Also bonus fixing of ofi_* includes from <> to ""
