+++ include/ofi.h
@@ -269,6 +269,26 @@ void ofi_strncatf(char *dest, size_t n, const char *fmt, ...);
 
 const char *ofi_hex_str(const uint8_t *data, size_t len);
 
+#define MAX_IPC_HANDLE_SIZE	64
+
+/*
+ * This structure is part of the
+ * the shm communication protocol
+ * defined in ofi_shm.h.
+ * Please make sure the SMR_VERSION are
+ * bumped and SMR_CMD_SIZE are large
+ * enough, for any changes in this
+ * structure.
+ */
+struct ipc_info {
+	uint64_t	iface;
+	uint64_t	base_addr;
+	uint64_t	base_length;
+	uint64_t	device;
+	uint64_t	offset;
+	uint8_t		ipc_handle[MAX_IPC_HANDLE_SIZE];
+};
+
 static inline uint64_t roundup_power_of_two(uint64_t n)
 {
 	if (!n || !(n & (n - 1)))
